+++ public/padachone.worker.js
@@ -1,5 +1,5 @@
-self.addEventListener("message", e => {
-  console.log(`Iam in receipt of your message : ${e.data}`);
+self.addEventListener("message", async (e) => {
+  console.log(`Iam in receipt of your message : `, JSON.stringify(e.data));
   const { type, msg } = e.data;
   let message;
   switch (type) {
@@ -8,9 +8,12 @@ self.addEventListener("message", e => {
       message = `${msg.action} : ${msg.message}`;
       sendWhatsappLogs(message, getbackToSite);
       break;
-    case "hooks":
-      message = `${msg.message}`;
-      sendWhatsappLogs(message, getbackToSite);
+    case "apod":
+      const apodJson = await fetchApod();
+      if (apodJson.url !== msg.current) {
+        getbackToSite({...apodJson, targetcomp : 'AppPages'});
+      }
+      // sendWhatsappLogs(message, getbackToSite);
       break;
     default:
       break;
@@ -35,4 +38,17 @@ async function sendWhatsappLogs(message, callback) {
   const respnse = await result.json();
   console.log(respnse);
   callback(`Successfully sent the following message : ${message}`);
+}
+
+async function fetchApod() {
+  const result = await fetch(
+    `https://api.nasa.gov/planetary/apod?api_key=DEMO_KEY`,
+    {
+      headers: {
+        Accept: "application/json"
+      }
+    }
+  );
+  const json = await result.json();
+  return json;
 }